+++ miso-web/src/main/webapp/scripts/bulk_libraryaliquot.js
@@ -172,7 +172,7 @@ BulkTarget.libraryaliquot = (function($) {
             required: designCode ? designCode.targetedSequencingRequired : false
           });
         }
-      }, BulkUtils.columns.librarySize);
+      }, BulkUtils.columns.qcPassed(true), BulkUtils.columns.librarySize);
 
       columns = columns.concat(BulkUtils.columns.concentration());
       columns = columns.concat(BulkUtils.columns.volume(false, config));