+++ miso-web/src/main/webapp/scripts/list_libraryaliquot.js
@@ -117,6 +117,12 @@ ListTarget.libraryaliquot = {
       "mRender": ListUtils.render.naIfNull,
       "iSortPriority": 0
     }, {
+      "sTitle": "QC",
+      "mData": "qcPassed",
+      "include": true,
+      "iSortPriority": 0,
+      "mRender": ListUtils.render.booleanChecks
+    }, {
       "sTitle": "Design",
       "mData": "libraryDesignCodeId",
       "include": Constants.isDetailedSample,