+++ sqlstore/src/main/resources/db/migration/V8505__library_aliquot_qc.sql
@@ -1,0 +1,1 @@
+ALTER TABLE LibraryAliquot ADD COLUMN qcPassed BOOLEAN;